Started on commenting and organizing the filestructure
