+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot/SoftwareTestingBot.java
@@ -4,10 +4,12 @@ import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.util.ElapsedTime;
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.ExampleCode.ExampleAction;
+import org.firstinspires.ftc.teamcode.SWEEP.Classes.SWEEPRobot;
+import org.firstinspires.ftc.teamcode.SWEEP.Classes.Subsystem;
 import org.firstinspires.ftc.teamcode.Robot.Subsystems.Drivetrain;
 import org.firstinspires.ftc.teamcode.SWEEP.Classes.Coordinate;
 
-public class SoftwareTestingBot extends SWEEPRobot{
+public class SoftwareTestingBot extends SWEEPRobot {
     // Robot subsystems
     Drivetrain drivetrain;
 
@@ -17,7 +19,7 @@ public class SoftwareTestingBot extends SWEEPRobot{
         this.drivetrain = new Drivetrain(hardwareMap, telemetry, false);
         subsystems = new Subsystem[]{
                 drivetrain,
-        };
+        };mbots
     }
 
     @Override
@@ -45,7 +47,7 @@ public class SoftwareTestingBot extends SWEEPRobot{
         return new ExampleAction(this, coordinate, time, range);
     }
     public ExampleAction exampleAction(){
-        return new ExampleAction(this, null, defaultActionTime, defaultActionRange);
+        return new ExampleAction(this, null, getDefaultActionTime(), getDefaultActionRange());
     }
 
 }
