+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot/Subsystems/Drivetrain.java
@@ -1,12 +1,11 @@
 package org.firstinspires.ftc.teamcode.Robot.Subsystems;
 
-import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.teamcode.Robot.Subsystem;
+import org.firstinspires.ftc.teamcode.SWEEP.Classes.Subsystem;
 //TODO: Finish commenting the class
 public class Drivetrain implements Subsystem {
     private DcMotor FLMotor;
