+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/SWEEP/Classes/SWEEPAction.java
@@ -1,12 +1,5 @@
 package org.firstinspires.ftc.teamcode.SWEEP.Classes;
 
-import android.graphics.SweepGradient;
-
-import com.qualcomm.robotcore.util.ElapsedTime;
-
-import org.firstinspires.ftc.teamcode.Robot.Action;
-import org.firstinspires.ftc.teamcode.Robot.Robot;
-
 /**
  * A class used to safely manage a robot's mechanisms with season specific code to the
  * generalized SWEEP Library, where the actions can be called autonomously
